+++ src/app/child-dev-project/children/child-details/form-components/basic-info/basic-info.component.scss
