+++ src/app/child-dev-project/children/child-details/form-components/dropout-form/dropout-form.component.scss
