+++ src/app/child-dev-project/children/child-details/form-components/dropout-form/dropout-form.component.spec.ts
@@ -1,14 +1,14 @@
 import { async, ComponentFixture, TestBed } from "@angular/core/testing";
 
 import { DropoutFormComponent } from "./dropout-form.component";
-import { EntityMapperService } from "../../../../core/entity/entity-mapper.service";
-import { Database } from "../../../../core/database/database";
-import { MockDatabase } from "../../../../core/database/mock-database";
-import { EntitySchemaService } from "../../../../core/entity/schema/entity-schema.service";
+import { EntityMapperService } from "../../../../../core/entity/entity-mapper.service";
+import { Database } from "../../../../../core/database/database";
+import { MockDatabase } from "../../../../../core/database/mock-database";
+import { EntitySchemaService } from "../../../../../core/entity/schema/entity-schema.service";
 import { MatFormFieldModule } from "@angular/material/form-field";
 import { MatDatepickerModule } from "@angular/material/datepicker";
 import { MatNativeDateModule } from "@angular/material/core";
-import { Child } from "../../model/child";
+import { Child } from "../../../model/child";
 import { MatInputModule } from "@angular/material/input";
 import { NoopAnimationsModule } from "@angular/platform-browser/animations";
 import { FormsModule } from "@angular/forms";
