+++ src/app/child-dev-project/children/child-details/form-components/health-form/health-form.component.spec.ts
@@ -1,16 +1,16 @@
 import { async, ComponentFixture, TestBed } from "@angular/core/testing";
 
 import { HealthFormComponent } from "./health-form.component";
-import { EntityMapperService } from "../../../../core/entity/entity-mapper.service";
-import { Database } from "../../../../core/database/database";
-import { MockDatabase } from "../../../../core/database/mock-database";
-import { EntitySchemaService } from "../../../../core/entity/schema/entity-schema.service";
+import { EntityMapperService } from "../../../../../core/entity/entity-mapper.service";
+import { Database } from "../../../../../core/database/database";
+import { MockDatabase } from "../../../../../core/database/mock-database";
+import { EntitySchemaService } from "../../../../../core/entity/schema/entity-schema.service";
 import { MatFormFieldModule } from "@angular/material/form-field";
 import { MatDatepickerModule } from "@angular/material/datepicker";
 import { MatNativeDateModule, MatOptionModule } from "@angular/material/core";
-import { Child } from "../../model/child";
-import { MatInput, MatInputModule } from "@angular/material/input";
-import { MatSelect, MatSelectModule } from "@angular/material/select";
+import { Child } from "../../../model/child";
+import { MatInputModule } from "@angular/material/input";
+import { MatSelectModule } from "@angular/material/select";
 import { NoopAnimationsModule } from "@angular/platform-browser/animations";
 import { FormsModule } from "@angular/forms";
